refactor(coordinator): extract QA/chat-answer session lifecycle into coordinator/qa_session.rs — behavior unchanged, 446 tests green
Pure move of 12 async/free functions (finalize_dictation_as_qa_question, begin/end/
cancel_qa_session, overlay transcription, answer_chat_dispatch, etc.) from
coordinator.rs into a new sibling submodule. Made pub(super); reached from the parent
and sibling 'qa' submodule via 'use qa_session::*;'. No logic/signature/public-API
changes; coordinator.rs 4298→3081 lines.

cargo check: 0 errors, 87 warnings (unchanged) · clippy: 0 errors
cargo test --lib: 446 passed · tsc / npm run build: green
